refactor: delete dead POST /api/learning-plan path
The standalone learning-plan endpoint had zero non-generated call
sites from the frontend — `learningPlanApiLearningPlanPost` was
unused, and `ModuleCard` (its only type consumer) was an orphan
component. The codebase also carried three incompatible classes
named `LearningPlan*` across `graph/state.py`,
`models/learning_plan.py`, and `models/assessment_api.py`, plus a
duplicate prompt and a duplicate LLM output schema — all serving a
code path nobody used.

This deletes Path 2 entirely. The assessment pipeline's
`generate_plan` terminal node remains the single learning-plan
generator, and the naming collision dissolves automatically.

Closes #165
