+++ backend/app/agents/schemas.py
@@ -165,43 +165,3 @@ class GapAnalysisOutput(BaseModel):
     gaps: list[GapAnalysisItemOutput] = Field(
         description="Skill gaps sorted by priority (critical first) then by gap size (largest first)"
     )
-
-
-# --- Standalone Learning Plan (used by /api/learning-plan route) ---
-
-
-class LearningPlanModuleOutput(BaseModel):
-    """A single learning module in a standalone learning plan phase."""
-
-    id: str = Field(description="Unique module identifier (e.g. mod-1)")
-    title: str = Field(description="Module title")
-    description: str = Field(description="What the learner will do")
-    type: Literal["theory", "quiz", "lab"] = Field(
-        description='Module type: "theory" (reading/video), "quiz" (practice problems), "lab" (hands-on project)'
-    )
-    phase: int = Field(description="Phase number this module belongs to")
-    skill_ids: list[str] = Field(description="Skill IDs addressed by this module")
-    duration_hours: int = Field(description="Estimated hours to complete")
-    objectives: list[str] = Field(description="Specific learning objectives")
-    resources: list[str] = Field(description="Learning resources (docs, courses, books)")
-
-
-class LearningPlanPhaseOutput(BaseModel):
-    """A single phase in a standalone learning plan."""
-
-    phase: int = Field(description="Phase order number starting from 1")
-    name: str = Field(description="Phase name")
-    description: str = Field(description="Phase description")
-    modules: list[LearningPlanModuleOutput] = Field(description="Modules in this phase")
-
-
-class LearningPlanOutput(BaseModel):
-    """Output schema for standalone learning plan generation."""
-
-    title: str = Field(description="Learning plan title")
-    summary: str = Field(description="2-3 sentence plan overview")
-    total_hours: int = Field(description="Total estimated hours for the entire plan (40-80)")
-    total_weeks: int = Field(description="Total estimated weeks (4-12)")
-    phases: list[LearningPlanPhaseOutput] = Field(
-        description="3-4 phases ordered by dependency and priority"
-    )
