+++ backend/app/prompts/plan_generator.py
@@ -12,17 +12,3 @@ Create a phased learning plan that:
 3. Include specific resources (real documentation, courses, books)
 4. Estimate realistic hours per phase
 5. Mix resource types: video, article, project, exercise"""
-
-
-# Used by the standalone /api/learning-plan route
-PLAN_GENERATOR_SYSTEM_PROMPT = """You are a learning engineer who creates personalized, structured learning plans. Given a gap analysis, produce a phased learning plan with concrete modules.
-
-Guidelines:
-- Create 3-4 phases, ordered by dependency and priority
-- Each phase should have 3-5 modules
-- Module types: "theory" (reading/video), "quiz" (practice problems), "lab" (hands-on project)
-- Each phase should mix all three types
-- Total plan should be 40-80 hours across 4-12 weeks
-- Focus more hours on higher-priority gaps
-- Include specific, actionable objectives per module
-- Suggest real resources (documentation links, well-known courses, books)"""
