+++ backend/tests/test_anthropic_error_handling.py
@@ -201,111 +201,6 @@ class TestGapAnalysisAnthropicErrors:
         assert "overloaded" in response.json()["detail"].lower()
 
 
-class TestLearningPlanAnthropicErrors:
-    """Test that Anthropic errors in learning-plan route are handled globally."""
-
-    @pytest.mark.asyncio
-    async def test_authentication_error_returns_401(self, setup_db):
-        with patch(
-            "app.routes.learning_plan.ainvoke_structured",
-            new_callable=AsyncMock,
-            side_effect=_make_auth_error(),
-        ):
-            async with AsyncClient(
-                transport=ASGITransport(app=_test_app), base_url="http://test"
-            ) as client:
-                response = await client.post(
-                    "/api/learning-plan",
-                    json={
-                        "gapAnalysis": {
-                            "overallReadiness": 50,
-                            "summary": "Needs work",
-                            "gaps": [
-                                {
-                                    "skillId": "ts",
-                                    "skillName": "TypeScript",
-                                    "currentLevel": 30,
-                                    "targetLevel": 80,
-                                    "gap": 50,
-                                    "priority": "high",
-                                    "recommendation": "Study",
-                                }
-                            ],
-                        }
-                    },
-                )
-
-        assert response.status_code == 401
-
-    @pytest.mark.asyncio
-    async def test_timeout_error_returns_504(self, setup_db):
-        with patch(
-            "app.routes.learning_plan.ainvoke_structured",
-            new_callable=AsyncMock,
-            side_effect=_make_timeout_error(),
-        ):
-            async with AsyncClient(
-                transport=ASGITransport(app=_test_app), base_url="http://test"
-            ) as client:
-                response = await client.post(
-                    "/api/learning-plan",
-                    json={
-                        "gapAnalysis": {
-                            "overallReadiness": 50,
-                            "summary": "Needs work",
-                            "gaps": [
-                                {
-                                    "skillId": "ts",
-                                    "skillName": "TypeScript",
-                                    "currentLevel": 30,
-                                    "targetLevel": 80,
-                                    "gap": 50,
-                                    "priority": "high",
-                                    "recommendation": "Study",
-                                }
-                            ],
-                        }
-                    },
-                )
-
-        assert response.status_code == 504
-        assert "timed out" in response.json()["detail"]
-
-    @pytest.mark.asyncio
-    async def test_internal_server_error_returns_502(self, setup_db):
-        with patch(
-            "app.routes.learning_plan.ainvoke_structured",
-            new_callable=AsyncMock,
-            side_effect=_make_internal_server_error(),
-        ):
-            async with AsyncClient(
-                transport=ASGITransport(app=_test_app), base_url="http://test"
-            ) as client:
-                response = await client.post(
-                    "/api/learning-plan",
-                    json={
-                        "gapAnalysis": {
-                            "overallReadiness": 50,
-                            "summary": "Needs work",
-                            "gaps": [
-                                {
-                                    "skillId": "ts",
-                                    "skillName": "TypeScript",
-                                    "currentLevel": 30,
-                                    "targetLevel": 80,
-                                    "gap": 50,
-                                    "priority": "high",
-                                    "recommendation": "Study",
-                                }
-                            ],
-                        }
-                    },
-                )
-
-        assert response.status_code == 502
-        assert "encountered an error" in response.json()["detail"]
-
-
 class TestAssessmentSSEAnthropicErrors:
     """Test that Anthropic errors in SSE stream produce structured [ERROR] events."""
