+++ docs/guides/api-reference.md
@@ -725,64 +725,6 @@ Priority levels: `critical` (gap > 40), `high` (gap > 25), `medium` (gap > 10), 
 
 ---
 
-### POST `/api/learning-plan`
-
-> **Requires authentication.** Returns 401 without a valid JWT cookie.
->
-> **Requires API key.** Returns 400 if the user has not configured an Anthropic API key.
-
-Generate a personalized learning plan from gap analysis.
-
-**Request**: `LearningPlanRequest`
-
-```json
-{
-  "gapAnalysis": {
-    "overallReadiness": 72,
-    "summary": "...",
-    "gaps": [...]
-  }
-}
-```
-
-**Response**: `LearningPlan`
-
-```json
-{
-  "summary": "A 6-week plan targeting distributed systems and security gaps.",
-  "totalHours": 48,
-  "phases": [
-    {
-      "phaseNumber": 1,
-      "title": "Foundations",
-      "concepts": ["microservices", "http_fundamentals"],
-      "rationale": "Build foundational understanding before tackling distributed systems.",
-      "resources": [
-        {
-          "type": "article",
-          "title": "Microservices Fundamentals",
-          "url": "https://microservices.io/patterns"
-        },
-        {
-          "type": "video",
-          "title": "HTTP Deep Dive",
-          "url": null
-        }
-      ],
-      "estimatedHours": 12
-    }
-  ]
-}
-```
-
-**Response** (400 — no API key configured):
-
-```json
-{"detail": "No API key configured. Please add your Anthropic API key in Settings."}
-```
-
----
-
 ### GET `/api/materials/{session_id}`
 
 > **Requires authentication.** Returns 401 without a valid JWT cookie.
