+++ frontend/src/lib/types.ts
@@ -4,9 +4,6 @@ export type {
   ProficiencyScore,
   GapItem,
   GapAnalysis,
-  LearningModule,
-  LearningPlan,
-  Phase,
   RoleSummary,
   RoleDetail,
   RoleLevelSummary,
